+++ daft/datatype.py
@@ -1203,29 +1203,6 @@ _EXT_TYPE_REGISTERED = False
 _STATIC_DAFT_EXTENSION: pa.ExtensionType | None = None
 
 
-class DaftExtension(pa.ExtensionType):  # type: ignore[misc]
-    def __init__(self, dtype: pa.DataType, metadata: bytes = b"") -> None:
-        # attributes need to be set first before calling
-        # super init (as that calls serialize)
-        self._metadata = metadata
-        super().__init__(dtype, "daft.super_extension")
-
-    def __reduce__(
-        self,
-    ) -> tuple[Callable[[pa.DataType, bytes], DaftExtension], tuple[pa.DataType, bytes]]:
-        return type(self).__arrow_ext_deserialize__, (self.storage_type, self.__arrow_ext_serialize__())
-
-    def __arrow_ext_serialize__(self) -> bytes:
-        return self._metadata
-
-    @classmethod
-    def __arrow_ext_deserialize__(cls, storage_type: pa.DataType, serialized: bytes) -> DaftExtension:
-        return cls(storage_type, serialized)
-
-    def __arrow_ext_equals__(self, other: pa.ExtensionType) -> bool:
-        return self.storage_type == other.storage_type and self._metadata == other._metadata
-
-
 def _ensure_registered_super_ext_type() -> None:
     global _EXT_TYPE_REGISTERED
     global _STATIC_DAFT_EXTENSION
@@ -1235,6 +1212,8 @@ def _ensure_registered_super_ext_type() -> None:
     if not _EXT_TYPE_REGISTERED:
         with _EXT_TYPE_REGISTRATION_LOCK:
             if not _EXT_TYPE_REGISTERED:
+                from daft.extension_type import DaftExtension
+
                 _STATIC_DAFT_EXTENSION = DaftExtension
                 pa.register_extension_type(DaftExtension(pa.null()))
                 import atexit
