+++ daft/expressions/expressions.py
@@ -38,12 +38,13 @@ from daft.daft import time_lit as _time_lit
 from daft.daft import timestamp_lit as _timestamp_lit
 from daft.daft import udf as _udf
 from daft.datatype import DataType, DataTypeLike, TimeUnit
-from daft.dependencies import pa, pc
+from daft.dependencies import pa
 from daft.expressions.testing import expr_structurally_equal
 from daft.logical.schema import Field, Schema
 from daft.series import Series, item_to_series
 
 if TYPE_CHECKING:
+    from daft.dependencies import pc
     from daft.io import IOConfig
     from daft.udf.legacy import BoundUDFArgs, InitArgsType, UninitializedUdf
     from daft.window import Window
@@ -380,7 +381,7 @@ class Expression:
 
     def to_arrow_expr(self) -> pc.Expression:
         """Returns this expression as a pyarrow.compute.Expression for integrations with other systems."""
-        from daft.expressions.visitor import _PyArrowExpressionVisitor
+        from daft.expressions.pyarrow_visitor import _PyArrowExpressionVisitor
 
         return _PyArrowExpressionVisitor().visit(self)
